[Changed] Heuristic-function change to return floating point.

+++ Hal/Heuristics/Dent.cpp
@@ -2,7 +2,7 @@
 #include "Dent.h"
 #include "../../Utility/CLOCKWISE_FOR.hpp"
 
-unsigned Dent::Execution(const Field& field){
+double Dent::Execution(const Field& field){
 	int score = 0;
 	
 	for(int i=0;i<FIELD_HEIGHT;i++){
